fix: renamed Webhooks test folder to EventDestinations
