+++ tests/Sinch.Tests/Conversation/EventDestinations/HmacAuthenticationValidationTests.cs
@@ -4,7 +4,7 @@ using FluentAssertions;
 using Sinch.Conversation.EventDestinations;
 using Xunit;
 
-namespace Sinch.Tests.Conversation.Webhooks
+namespace Sinch.Tests.Conversation.EventDestinations
 {
     public class HmacAuthenticationValidationTests
     {
